refactor(app): split public app into spec file layout
Reorganize the consolidated public-app UI into the SPEC.md structure (no behavior or
visual change — verified by build + screenshots):
- components/wf/* — primitives split into Pin, Avatar, SearchPill, Handle, ViewToggle,
  source-meta, with an index barrel (consumers keep importing '@/components/wf').
- components/desktop/{Sidebar,SourceRail} — sidebar + source rail extracted from page.
- components/mobile/PeekSheet — the 3-snap bottom sheet extracted from page; page.tsx is
  now a thin AppShell that owns data and composes Sidebar/PeekSheet.
- types/view-models.ts — VM aliases over the DTOs (Bb-4).
- app/dev/page.tsx — PlaceCard states sandbox (Ba-5).
