+++ sunrei-frontend/packages/sunrei-app/src/components/wf/Handle.tsx
@@ -1,0 +1,6 @@
+'use client';
+
+/** Sheet grab handle. */
+export function Handle() {
+  return <div className="mx-auto mt-2 mb-1.5 h-1 w-9 rounded-full bg-line2" />;
+}
