+++ sunrei-frontend/packages/sunrei-app/src/components/wf/index.ts
@@ -1,0 +1,8 @@
+// Wireframe primitives (Sunrei Wireframes.html). Cornflower-blue accent, cream/greige
+// neutrals, teardrop pins, avatar-initial source rows. Tokens live in globals.css.
+export { Pin } from './Pin';
+export { Avatar, AvatarCluster } from './Avatar';
+export { SearchPill } from './SearchPill';
+export { Handle } from './Handle';
+export { ViewToggle } from './ViewToggle';
+export { sourceVerb } from './source-meta';
